refactor `src/view` file structure
