+++ src/view/menu/InfoBar.svelte
