+++ src/view/menu/TickSpinner.svelte
