+++ src/view/panel/IdleCallbackRequestHistory.svelte
@@ -1,61 +1,71 @@
 <script lang="ts">
-  import type {
-    TCancelAnimationFrameHistory,
-    TRequestAnimationFrameHistory,
-  } from '../../wrapper/AnimationWrapper.ts';
   import {
-    DEFAULT_SORT_RAF,
+    RicFact,
+    type TCancelIdleCallbackHistory,
+    type TRequestIdleCallbackHistory,
+  } from '../../wrapper/IdleWrapper.ts';
+  import {
+    DEFAULT_SORT_RIC,
     ESortOrder,
     getSettings,
     setSettings,
   } from '../../api/settings.ts';
   import { compareByFieldOrder } from '../../api/comparator.ts';
-  import Variable from './Variable.svelte';
-  import SortableColumn from './SortableColumn.svelte';
-  import FrameSensitiveTime from './FrameSensitiveTime.svelte';
-  import TraceBreakpoint from './TraceBreakpoint.svelte';
-  import Dialog from './Dialog.svelte';
-  import Alert from './Alert.svelte';
-  import AnimationCancelHistory from './AnimationCancelHistory.svelte';
-  import TraceBypass from './TraceBypass.svelte';
-  import CancelableCallMetric from './CancelableCallMetric.svelte';
-  import CallstackCell from './CallstackCell.svelte';
+  import { msToHms } from '../../api/time.ts';
+  import Variable from '../components/Variable.svelte';
+  import IdleCallbackCancelHistory from './IdleCallbackCancelHistory.svelte';
+  import Dialog from '../components/Dialog.svelte';
+  import Alert from '../components/Alert.svelte';
+  import SortableColumn from './components/SortableColumn.svelte';
+  import FrameSensitiveTime from './components/FrameSensitiveTime.svelte';
+  import TraceBreakpoint from './components/TraceBreakpoint.svelte';
+  import TraceBypass from './components/TraceBypass.svelte';
+  import CancelableCallMetric from './components/CancelableCallMetric.svelte';
+  import type { TFactsMap } from '../../wrapper/Fact.ts';
+  import FactsCell from './components/FactsCell.svelte';
+  import CallstackCell from './components/CallstackCell.svelte';
 
   let {
-    rafHistory,
-    cafHistory,
+    ricHistory,
+    cicHistory = null,
     caption = '',
   }: {
-    rafHistory: TRequestAnimationFrameHistory[];
-    cafHistory: TCancelAnimationFrameHistory[] | null;
+    ricHistory: TRequestIdleCallbackHistory[];
+    cicHistory: TCancelIdleCallbackHistory[] | null;
     caption: string;
   } = $props();
-  let sortField = $state(DEFAULT_SORT_RAF.field);
-  let sortOrder = $state(DEFAULT_SORT_RAF.order);
+  let sortField = $state(DEFAULT_SORT_RIC.field);
+  let sortOrder = $state(DEFAULT_SORT_RIC.order);
   let dialogEl: Dialog | null = null;
   let alertEl: Alert | null = null;
   let sortedMetrics = $derived.by(() =>
-    rafHistory.toSorted(compareByFieldOrder(sortField, sortOrder))
+    ricHistory.toSorted(compareByFieldOrder(sortField, sortOrder))
   );
+  const RicFacts: TFactsMap = new Map([
+    [RicFact.BAD_DELAY, {
+      tag: 'D',
+      details: 'Delay is not a positive number or undefined',
+    }],
+  ]);
 
   getSettings().then((settings) => {
-    sortField = settings.sortRequestAnimationFrame.field;
-    sortOrder = settings.sortRequestAnimationFrame.order;
+    sortField = settings.sortRequestIdleCallback.field;
+    sortOrder = settings.sortRequestIdleCallback.order;
   });
 
   function onChangeSort(_field: string, _order: ESortOrder) {
-    sortField = <keyof TRequestAnimationFrameHistory> _field;
+    sortField = <keyof TRequestIdleCallbackHistory> _field;
     sortOrder = _order;
 
     setSettings({
-      sortRequestAnimationFrame: {
+      sortRequestIdleCallback: {
         field: $state.snapshot(sortField),
         order: $state.snapshot(sortOrder),
       },
     });
   }
 
-  let cafHistoryMetrics: TCancelAnimationFrameHistory[] = $state([]);
+  let cicHistoryMetrics: TCancelIdleCallbackHistory[] = $state([]);
 
   function onFindRegressors(regressors: string[] | null) {
     if (!dialogEl || !alertEl || !regressors?.length) {
@@ -64,13 +74,13 @@
 
     for (let n = regressors.length - 1; n >= 0; n--) {
       const traceId = regressors[n];
-      let record = cafHistory?.find((r) => r.traceId === traceId);
+      let record = cicHistory?.find((r) => r.traceId === traceId);
       if (record) {
-        cafHistoryMetrics.push(record);
+        cicHistoryMetrics.push(record);
       }
     }
 
-    if (cafHistoryMetrics.length) {
+    if (cicHistoryMetrics.length) {
       dialogEl.show();
     } else {
       alertEl.show();
@@ -78,32 +88,33 @@
   }
 
   function onCloseDialog() {
-    cafHistoryMetrics.splice(0);
+    cicHistoryMetrics.splice(0);
   }
 </script>
 
 <Dialog
   bind:this={dialogEl}
   eventClose={onCloseDialog}
-  title="Places from which requestAnimationFrame with current callstack was prematurely canceled"
-  description="The information is actual only on time of demand. Requires cancelAnimationFrame panel enabled."
+  title="Places from which requestIdleCallback with current callstack was prematurely canceled"
+  description="The information is actual only on time of demand. Requires cancelIdleCallback panel enabled."
 >
-  <AnimationCancelHistory
+  <IdleCallbackCancelHistory
     caption="Canceled by"
-    cafHistory={$state.snapshot(cafHistoryMetrics)}
+    cicHistory={$state.snapshot(cicHistoryMetrics)}
   />
 </Dialog>
 
 <Alert bind:this={alertEl} title="Attention">
-  Requires cancelAnimationFrame panel enabled
+  Requires cancelIdleCallback panel enabled
 </Alert>
 
 <table data-navigation-tag={caption}>
   <thead class="sticky-header">
     <tr>
       <th class="w-full">
-        {caption} Callstack [<Variable value={rafHistory.length} />]
+        {caption} Callstack [<Variable value={ricHistory.length} />]
       </th>
+      <th class="ta-c">didTimeout</th>
       <th class="ta-r">
         <SortableColumn
           field="selfTime"
@@ -112,7 +123,14 @@
           eventChangeSorting={onChangeSort}
         >Self</SortableColumn>
       </th>
-      <th class="ta-c" title="Calls per second">CPS</th>
+      <th class="ta-c">
+        <SortableColumn
+          field="facts"
+          currentField={sortField}
+          currentFieldOrder={sortOrder}
+          eventChangeSorting={onChangeSort}
+        ><span class="icon -facts"></span></SortableColumn>
+      </th>
       <th class="ta-c">
         <SortableColumn
           field="calls"
@@ -128,6 +146,14 @@
           currentFieldOrder={sortOrder}
           eventChangeSorting={onChangeSort}
         >Handler</SortableColumn>
+      </th>
+      <th class="ta-r">
+        <SortableColumn
+          field="delay"
+          currentField={sortField}
+          currentFieldOrder={sortOrder}
+          eventChangeSorting={onChangeSort}
+        >Delay</SortableColumn>
       </th>
       <th>
         <SortableColumn
@@ -151,8 +177,11 @@
             traceDomain={metric.traceDomain}
           />
         </td>
+        <td class="ta-c">{metric.didTimeout}</td>
         <td class="ta-r"><FrameSensitiveTime value={metric.selfTime} /></td>
-        <td class="ta-c">{metric.cps || undefined}</td>
+        <td class="ta-c">
+          <FactsCell facts={metric.facts} factsMap={RicFacts} />
+        </td>
         <td class="ta-c">
           <CancelableCallMetric
             calls={metric.calls}
@@ -162,6 +191,7 @@
           />
         </td>
         <td class="ta-c"><Variable value={metric.handler} /></td>
+        <td class="ta-r" title={msToHms(metric.delay)}>{metric.delay}</td>
         <td class="ta-r">
           {#if metric.online}
             <Variable value={metric.online} />
