+++ src/view/panel/TimersClearHistory.svelte
@@ -1,47 +1,38 @@
 <script lang="ts">
-  import type {
-    TClearTimerHistory,
-    TSetTimerHistory,
-  } from '../../wrapper/TimerWrapper.ts';
+  import type { TClearTimerHistory } from '../../wrapper/TimerWrapper.ts';
   import {
-    DEFAULT_SORT_SET_TIMERS,
-    ESortOrder,
+    DEFAULT_SORT_CLEAR_TIMERS,
+    type ESortOrder,
     getSettings,
     setSettings,
   } from '../../api/settings.ts';
   import { compareByFieldOrder } from '../../api/comparator.ts';
-  import Variable from './Variable.svelte';
-  import SortableColumn from './SortableColumn.svelte';
-  import TimersSetHistoryMetric from './TimersSetHistoryMetric.svelte';
+  import Variable from '../components/Variable.svelte';
+  import SortableColumn from './components/SortableColumn.svelte';
+  import TimersClearHistoryMetric from './components/TimersClearHistoryMetric.svelte';
 
   let {
-    setTimerHistory,
-    clearTimeoutHistory,
-    clearIntervalHistory,
+    clearTimerHistory,
     caption,
-  }: {
-    setTimerHistory: TSetTimerHistory[];
-    clearTimeoutHistory: TClearTimerHistory[] | null;
-    clearIntervalHistory: TClearTimerHistory[] | null;
-    caption?: string;
-  } = $props();
-  let sortField = $state(DEFAULT_SORT_SET_TIMERS.field);
-  let sortOrder = $state(DEFAULT_SORT_SET_TIMERS.order);
+  }: { clearTimerHistory: TClearTimerHistory[]; caption: string } =
+    $props();
+  let sortField = $state(DEFAULT_SORT_CLEAR_TIMERS.field);
+  let sortOrder = $state(DEFAULT_SORT_CLEAR_TIMERS.order);
   let sortedMetrics = $derived.by(() =>
-    setTimerHistory.toSorted(compareByFieldOrder(sortField, sortOrder))
+    clearTimerHistory.toSorted(compareByFieldOrder(sortField, sortOrder))
   );
 
   getSettings().then((settings) => {
-    sortField = settings.sortSetTimers.field;
-    sortOrder = settings.sortSetTimers.order;
+    sortField = settings.sortClearTimers.field;
+    sortOrder = settings.sortClearTimers.order;
   });
 
   function onChangeSort(_field: string, _order: ESortOrder) {
-    sortField = <keyof TSetTimerHistory> _field;
+    sortField = <keyof TClearTimerHistory> _field;
     sortOrder = _order;
 
     setSettings({
-      sortSetTimers: {
+      sortClearTimers: {
         field: $state.snapshot(sortField),
         order: $state.snapshot(sortOrder),
       },
@@ -53,15 +44,7 @@
   <thead class="sticky-header">
     <tr>
       <th class="w-full">
-        {caption} Callstack [<Variable value={setTimerHistory.length} />]
-      </th>
-      <th class="ta-r">
-        <SortableColumn
-          field="selfTime"
-          currentField={sortField}
-          currentFieldOrder={sortOrder}
-          eventChangeSorting={onChangeSort}
-        >Self</SortableColumn>
+        {caption} Callstack [<Variable value={clearTimerHistory.length} />]
       </th>
       <th class="ta-c">
         <SortableColumn
@@ -95,14 +78,6 @@
           eventChangeSorting={onChangeSort}
         >Delay</SortableColumn>
       </th>
-      <th>
-        <SortableColumn
-          field="online"
-          currentField={sortField}
-          currentFieldOrder={sortOrder}
-          eventChangeSorting={onChangeSort}
-        >Set</SortableColumn>
-      </th>
       <th title="Bypass"><span class="icon -bypass"></span></th>
       <th title="Breakpoint"><span class="icon -breakpoint"></span></th>
     </tr>
@@ -110,11 +85,7 @@
 
   <tbody>
     {#each sortedMetrics as metric (metric.traceId)}
-      <TimersSetHistoryMetric
-        {metric}
-        {clearTimeoutHistory}
-        {clearIntervalHistory}
-      />
+      <TimersClearHistoryMetric {metric} />
     {/each}
   </tbody>
 </table>
