+++ src/view/panel/components/TraceBypass.svelte
@@ -3,16 +3,16 @@
     getSettings,
     onSettingsChange,
     setSettings,
-  } from '../../api/settings.ts';
+  } from '../../../api/settings.ts';
 
   let { traceId }: { traceId: string } = $props();
-  let trace4Debug: string | null = $state.raw(null);
+  let trace4Bypass: string | null = $state.raw(null);
 
   getSettings().then((settings) => {
-    trace4Debug = settings.trace4Debug;
+    trace4Bypass = settings.trace4Bypass;
 
     onSettingsChange((settings) => {
-      trace4Debug = settings.trace4Debug;
+      trace4Bypass = settings.trace4Bypass;
     });
   });
 
@@ -20,7 +20,7 @@
     e.preventDefault();
 
     setSettings({
-      trace4Debug: trace4Debug === traceId
+      trace4Bypass: trace4Bypass === traceId
         ? null
         : $state.snapshot(traceId),
     });
@@ -30,11 +30,11 @@
 <div class="sensor">
   <a
     href="void(0)"
-    class:active={trace4Debug === traceId}
-    aria-label="Place breakpoint"
+    class:active={trace4Bypass === traceId}
+    aria-label="Place bypass"
     onclick={onToggle}
   >
-    <span class="icon -breakpoint"></span>
+    <span class="icon -bypass"></span>
   </a>
 </div>
